% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dg0i}
\alias{dg0i}
\title{The G0I Distribution}
\usage{
dg0i(x, alpha, gama, L)
}
\arguments{
\item{x}{vector of quantiles.}

\item{alpha}{value for alpha parameter, this value is < 0.}

\item{gama}{value for gamma parameter, this value is > 0.}

\item{L}{value for L parameter, this value is > 0.}
}
\value{
\code{dg0i} gives the density
}
\description{
\code{dg0i} function related to the G0I Distribution.
}
\examples{

set.seed(1) # determining a seed

alpha <- -1.5
gama <- 2
L <- 8

Xobs<- rg0i(100,alpha, gama, L)
dg0i(Xobs,alpha, gama, L)


}
\references{
\itemize{
\item Nascimento, Abraao DC, Renato J. Cintra, and Alejandro C. Frery. "Hypothesis testing in speckled data with stochastic distances." IEEE Transactions on geoscience and remote sensing 48.1 (2009): 373-385.
}
}
