% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancesg0i.R
\name{distancesg0i}
\alias{distancesg0i}
\title{G0I Distances}
\usage{
distancesg0i(par_1, par_2, p = 1/2, type = "KL")
}
\arguments{
\item{par_1}{parameters vector for sample 1}

\item{par_2}{parameters vector for sample 2}

\item{p}{parameter for XXXX}

\item{type}{type of distance for calculate}
}
\value{
Distance value between samples.
}
\description{
\code{distancesg0i} G0I Distances
}
\examples{

set.seed(1) # determining a seed

alpha <- -1.5
gama <- 2
L <- 8

distancesg0i(c(alpha,gama,L),c(alpha,gama,L))
res <- distancesg0i(c(alpha,gama,L),c(alpha,gama,4))
print(res)


}
\references{
\itemize{
\item Nascimento, Abraao DC, Renato J. Cintra, and Alejandro C. Frery. "Hypothesis testing in speckled data with stochastic distances." IEEE Transactions on geoscience and remote sensing 48.1 (2009): 373-385.
}
}
