% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancetable.R
\name{distancesg0itable}
\alias{distancesg0itable}
\title{G0I Distances Table}
\usage{
distancesg0itable(image, marks1, marks2, p = 1/2)
}
\arguments{
\item{image}{an array image with dimensions X,Y,Z}

\item{marks1}{vector 1 of marks with the follow order c(left, right, up, down)}

\item{marks2}{vector 2 of marks with the follow order c(left, right, up, down)}

\item{p}{parameter for XXXX}
}
\value{
Distances table between samples obtained of marks.
}
\description{
\code{distancesg0itable} G0I Distances Table
}
\examples{

set.seed(1) # determining a seed

data("SanFrancisco150")
image <- SanFrancisco150

# left,right, up,down)
marks1 <- c(10,30,50,100)
marks2 <- c(20,50,80,130)

teste <- distancesg0itable(image,marks1, marks2, p = 1/2)
teste


}
\references{
\itemize{
\item Nascimento, Abraao DC, Renato J. Cintra, and Alejandro C. Frery. "Hypothesis testing in speckled data with stochastic distances." IEEE Transactions on geoscience and remote sensing 48.1 (2009): 373-385.
}
}
