% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{g0i}
\alias{g0i}
\alias{dg0i}
\alias{pg0i}
\alias{qg0i}
\alias{rg0i}
\title{The G0I Distribution}
\usage{
dg0i(x, alpha, gamma, L)

rg0i(n, alpha, gamma, L)

pg0i(q, alpha, gamma, L, lower.tail = TRUE, log.p = FALSE)

qg0i(p, alpha, gamma, L, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{alpha}{value for alpha parameter, this value is < 0.}

\item{gamma}{value for gamma parameter, this value is > 0.}

\item{L}{value for L parameter, this value is > 0.}

\item{n}{Number of observations.}

\item{lower.tail}{Logical; if TRUE (default), probabilities are \eqn{P[X ≤ x]} otherwise, \eqn{P[X > x]}.}

\item{log.p}{Logical; if TRUE, probabilities p are given as \eqn{log(p)}.}

\item{p}{Vector of probabilities.}
}
\value{
The density, distribution function, quantile function or random generation for the G0I distribution.
}
\description{
Density, distribution function, quantile function and random generation
for the g0i distribution with parameters alpha, gamma and L.
}
\examples{
set.seed(1) # determining a seed

alpha <- -1.5
gama <- 2
L <- 8

Xobs<- rg0i(100,alpha, gama, L)
dg0i(Xobs,alpha, gama, L)
qg0i(Xobs,alpha, gama, L)

qobs<- runif(10)
pg0i(qobs,alpha, gama, L)

}
\references{
\itemize{
\item Nascimento, Abraao DC, Renato J. Cintra, and Alejandro C. Frery. "Hypothesis testing in speckled data with stochastic distances." IEEE Transactions on geoscience and remote sensing 48.1 (2009): 373-385.
}
}
