% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationg0i.R
\name{g0iestimation}
\alias{g0iestimation}
\title{G0I Estimation}
\usage{
g0iestimation(Theta, x, L_fixed = FALSE)
}
\arguments{
\item{Theta}{vector of random initial values considering the assumptions of the parameters, \eqn{alpha < 0}, \eqn{gamma > 0 } and \eqn{L > 0}.}

\item{x}{vector of observations.}

\item{L_fixed}{logical; if TRUE, the parameter \eqn{L} is fixed. \eqn{L} is not fixed by default, \eqn{L = FALSE}.}
}
\description{
\code{g0iestimation} Estimation of G0I distribution parameters
}
\examples{

set.seed(1) # determining a seed

alpha <- -1.5
gama <- 2
L <- 8

y <- rg0i(1000, alpha, gama, L)

# Estimation with all parameters
g0iestimation(c(-1, 1, 1), y)

# Estimation with fixed L
g0iestimation(c(-1, 1, 1), y, L_fixed = TRUE)

}
\references{
\itemize{
\item Nascimento, Abraao DC, Renato J. Cintra, and Alejandro C. Frery. "Hypothesis testing in speckled data with stochastic distances." IEEE Transactions on geoscience and remote sensing 48.1 (2009): 373-385.
}
}
