% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SanFrancisco150.R
\docType{data}
\name{SanFrancisco150}
\alias{SanFrancisco150}
\title{San Francisco Data}
\format{
A dataset with 150 rows, 150 columns and 6 dimensions.
}
\usage{
SanFrancisco150
}
\description{
A dataset containing training data from San Francisco
}
\keyword{datasets}
